Brain-tumor-class-nillion/quickstart/nada_quickstart_programs/requirements.txt
asttokens==2.4.1
nada-dsl==0.4.0
parsial==0.1.0
richreports==0.2.0
six==1.16.0
sortedcontainers==2.4.0
py-nillion-client
nillion-python-helpers
python-dotenv==1.0.0
pytest-asyncio>=0.23.6
cosmpy>=0.9.2
torchvision~=0.18.1
matplotlib~=3.9.0
pillow==10.3.0
streamlit
